+++ torchvision_models/segmentation/erfnet.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from collections import OrderedDict
-from ..common_models import SpatialConv, EDLaneExist
+from ..common_models import SpatialConv, EDLaneExist, non_bottleneck_1d
 
 
 class DownsamplerBlock(nn.Module):
@@ -20,37 +20,6 @@ class DownsamplerBlock(nn.Module):
         output = torch.cat([self.conv(input), self.pool(input)], 1)
         output = self.bn(output)
         return F.relu(output)
-    
-
-class non_bottleneck_1d(nn.Module):
-    def __init__(self, chann, dropprob, dilated):        
-        super().__init__()
-        self.conv3x1_1 = nn.Conv2d(chann, chann, (3, 1), stride=1, padding=(1, 0), bias=True)
-        self.conv1x3_1 = nn.Conv2d(chann, chann, (1, 3), stride=1, padding=(0, 1), bias=True)
-        self.bn1 = nn.BatchNorm2d(chann, eps=1e-03)
-        self.conv3x1_2 = nn.Conv2d(chann, chann, (3, 1), stride=1, padding=(1*dilated, 0),
-                                   bias=True, dilation=(dilated, 1))
-        self.conv1x3_2 = nn.Conv2d(chann, chann, (1, 3), stride=1, padding=(0, 1*dilated),
-                                   bias=True, dilation=(1, dilated))
-        self.bn2 = nn.BatchNorm2d(chann, eps=1e-03)
-        self.dropout = nn.Dropout2d(dropprob)
-
-    def forward(self, input):
-        output = self.conv3x1_1(input)
-        output = F.relu(output)
-        output = self.conv1x3_1(output)
-        output = self.bn1(output)
-        output = F.relu(output)
-
-        output = self.conv3x1_2(output)
-        output = F.relu(output)
-        output = self.conv1x3_2(output)
-        output = self.bn2(output)
-
-        if self.dropout.p != 0:
-            output = self.dropout(output)
-        
-        return F.relu(output + input)
 
 
 class Encoder(nn.Module):
